Rename test into tests
