+++ tests/unit/test_datafiles.py
@@ -2,16 +2,16 @@ import minplascalc as mpc
 
 
 def test_build_monatomic_species_json():
-    result = mpc.species.from_name('O+')
+    result = mpc.species.from_name("O+")
 
-    assert result.stoichiometry == {'O': 1}
+    assert result.stoichiometry == {"O": 1}
     assert result.chargenumber == 1
     assert result.sources is not None
 
 
 def test_build_diatomic_species_json():
-    result = mpc.species.from_name('CO')
-    
-    assert result.stoichiometry == {'C': 1, 'O': 1}
+    result = mpc.species.from_name("CO")
+
+    assert result.stoichiometry == {"C": 1, "O": 1}
     assert result.chargenumber == 0
     assert result.sources is not None
